# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  go-micro_pkg_httpx.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_refund.InitiateRequest:
    properties:
      order_id:
        type: string
      reason:
        type: string
      refund_id:
        type: string
      refund_type:
        type: string
    type: object
  internal_refund.RollbackRequest:
    properties:
      order_id:
        type: string
      reason:
        type: string
      refund_id:
        type: string
    type: object
  internal_refund.StatusRequest:
    properties:
      refund_id:
        type: string
    type: object
  internal_task.CreateTaskRequest:
    properties:
      biz_no:
        type: string
      order_id:
        type: string
      type:
        type: string
    type: object
  internal_user.CreateUserRequest:
    properties:
      password:
        type: string
      role:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Go-Micro 网关服务 API
  title: Go-Micro Gateway API
  version: "1.0"
paths:
...
  /api/v1/refund/initiate:
    post:
      consumes:
      - application/json
      parameters:
      - description: initiate refund
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.InitiateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Initiate refund
      tags:
      - Refund
  /api/v1/refund/rollback:
    post:
      consumes:
      - application/json
      parameters:
      - description: rollback refund
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.RollbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Rollback on cancel
      tags:
      - Refund
  /api/v1/refund/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: refund status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_refund.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: Refund status
      tags:
      - Refund
...
  /payments/{id}/refund:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 退款
      tags:
      - Payment
...
  /tasks:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_task.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建任务
      tags:
      - Task
  /tasks/{id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询任务
      tags:
      - Task
  /tasks/{id}/retry:
    post:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 手动重试任务
      tags:
      - Task
  /tasks/by-order/{order_id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 按订单查询任务
      tags:
      - Task
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建用户
      tags:
      - User
  /users/{id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询用户
      tags:
      - User
swagger: "2.0"