# internal/user/handler.go
package user

import (
	"github.com/gin-gonic/gin"
	"go-micro/pkg/errx"
	"go-micro/pkg/httpx"
)

type Handler struct {
	svc *Service
}

func NewHandler(svc *Service) *Handler {
	return &Handler{svc: svc}
}

func (h *Handler) Register(r *gin.Engine) {
	r.GET("/health", func(c *gin.Context) {
		code, body := httpx.OK(gin.H{"status": "ok"})
		c.JSON(code, body)
	})

	r.POST("/users", h.create)
	r.GET("/users/:id", h.get)
}

// @Summary 创建用户
// @Tags User
// @Accept json
// @Produce json
// @Param body body CreateUserRequest true "用户创建请求"
// @Success 200 {object} httpx.Response
// @Router /users [post]
func (h *Handler) create(c *gin.Context) {
	var req CreateUserRequest
	if err := c.ShouldBindJSON(&req); err != nil || req.Username == "" || req.Password == "" {
		code, body := httpx.Fail(errx.CodeInvalidRequest, errx.MsgInvalidRequest)
		c.JSON(code, body)
		return
	}
	user, err := h.svc.Create(req)
	if err != nil {
		code, body := httpx.Fail(errx.CodeInternalError, errx.MsgUserCreateFailed)
		c.JSON(code, body)
		return
	}
	code, body := httpx.OK(user)
	c.JSON(code, body)
}

// @Summary 查询用户
// @Tags User
// @Produce json
// @Param id path string true "用户ID"
// @Success 200 {object} httpx.Response
// @Router /users/{id} [get]
func (h *Handler) get(c *gin.Context) {
	id := c.Param("id")
	user, err := h.svc.Get(id)
	if err != nil {
		code, body := httpx.Fail(errx.CodeNotFound, errx.MsgUserNotFound)
		c.JSON(code, body)
		return
	}
	code, body := httpx.OK(user)
	c.JSON(code, body)
}